pyproject.toml
[project]
name = "project"
version = "0" # Controlled by environment variable
requires-python = "==3.14.*" # Controlled by .python-version
dependencies = [
    "frozendict>=2.4.7",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-zen>=0.16.0",
    "joblib>=1.5.3",
    "mlflow>=3.10.1",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "pandera[pandas]>=0.27.1",
    "pip>=25.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "scikit-learn>=1.8.0",
    "skops>=0.13.0",
    "sqlalchemy>=2.0.45",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "complexipy>=5.2.0",
    "dirty-equals>=0.11",
    "inline-snapshot>=0.32.2",
    "inline-snapshot-pandas>=0.2.0",
    "interrogate>=1.7.0",
    "prek>=0.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-profiling>=1.8.1",
    "pytest-raises>=0.11",
    "pytest-randomly>=4.0.1",
    "pytest-sugar>=1.1.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.14.10",
    "tach>=0.33.3",
    "ty>=0.0.8",
]
notebook = [
    "ipytest>=0.14.2",
    "ipython!=9.10.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "plotly>=6.5.2",
    "seaborn>=0.13.2",
]

[project.scripts]
train = "project.entrypoints.train:cli"
predict = "project.entrypoints.predict:cli"
evaluate = "project.entrypoints.evaluate:cli"

[tool.uv]
default-groups = [
    "dev",
    "notebook",
]
override-dependencies = [
    "optuna>=2.10.0",
    "sqlalchemy>=1.3.0",
]

[build-system]
requires = ["uv_build<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 88
extend-exclude = [
    "src/**/legacy", # Skip legacy code
]

[tool.ruff.lint]
extend-select = [
    "I",     # organize-imports
    "E501",  # line-too-long
    "UP006", # non-pep585-annotation
    "UP035", # deprecated-import
    "D200",  # unnecessary-multiline-docstring
    "D209",  # new-line-after-last-paragraph
    "D210",  # surrounding-whitespace
    "D212",  # multi-line-summary-first-line
    "D403",  # first-word-capitalized
    "D405",  # non-capitalized-section-name
    "D416",  # missing-section-name-colon
]
ignore = [
    "E731", # lambda-assignment
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/test_*.py" = ["S101"] # assert
"**/*.ipynb" = ["B018"]            # useless-expression

[tool.ruff.format]
line-ending = "lf"

[tool.ty.src]
exclude = [
    "src/**/legacy", # Skip legacy code
]

[tool.pytest]
norecursedirs = [
    ".*",          # pytest default
    "entrypoints", # Never test entrypoints!
    "legacy",      # Skip legacy code
]
timeout = "180" # Alert on hanging tests
strict_xfail = true # Error on successful tests that should fail
addopts = [
    "-ra",                     # More insights
    "-vv",                     # Show parametrized test names
    "-p",
    "no:tach",                 # Skipping tests skews coverage
    "--strict-markers",        # Prevent silent bugs
    "--strict-config",         # Report errors in config
    "--doctest-modules",       # Test code examples in doc strings
    "--cov",                   # Report coverage by default
    "--import-mode=importlib", # Prevent importing from local src
]
testpaths = ["tests", "src"] # Speed up collection; include src for doctest

[tool.coverage.run]
source = [
    "src",
    "tests", # Spot outdated tests
]
omit = [
    "notebooks/**",
    "src/**/entrypoints/**",
    "src/**/legacy/**",
]
branch = true

[tool.coverage.report]
fail_under = 100     # Clean coverage. Use `#pragma: no cover` to skip coverage.
ignore_errors = true
show_missing = true
skip_covered = true
skip_empty = true

[tool.inline-snapshot]
format-command = "uv run ruff check --fix-only --stdin-filename {filename} | uv run ruff format --stdin-filename {filename}"

[tool.interrogate]
omit-covered-files = true
verbose = 2
fail-under = 100
whitelist-regex = ["test_.*"]
exclude = ["*.cache"]
style = "google"

[tool.complexipy]
paths = ["src", "tests"] # Cognitive complexity: How intuitive is a function for humans.
max-complexity-allowed = 10
failed = true
exclude = [
    "project/data/legacy",        # No glob patterns supported
    "project/entrypoints/legacy",
    "project/features/legacy",
    "project/models/legacy",
    "project/util/legacy",
]
